move components to their own files
